requirements.txt
cloudinary==1.44.0
Flask==3.1.1
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
numpy==2.2.6
opencv_python==4.10.0.84
opencv_python_headless==4.8.1.78
Pillow==11.2.1
python-dotenv==1.1.0
torch==2.6.0
torchvision==0.21.0
Werkzeug==3.1.3
gunicorn
psutil